Внесение исправлений по замечаниям от наставника
Изменены названия компонентов-страниц.
Константы вынесены в отдельный файл.
