+++ src/const.tsx
@@ -1,0 +1,4 @@
+export const Setting = {
+  PlacesCount: 312,
+  PlacesOnPage: 5,
+};
