+++ src/index.tsx
@@ -2,16 +2,19 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
 
-const PLACES_COUNT = 5;
+import { Setting } from './const';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 
+const { PlacesCount, PlacesOnPage } = Setting;
+
 root.render(
   <React.StrictMode>
     <App
-      placesCount={PLACES_COUNT}
+      placesCount = { PlacesCount }
+      placesOnPage = { PlacesOnPage }
     />
   </React.StrictMode>
 );
