+++ src/pages/login/login.tsx
@@ -1,4 +1,4 @@
-function LoginScreen(): JSX.Element {
+function Login(): JSX.Element {
 
   return (
     <div className="page page--gray page--login">
@@ -43,4 +43,4 @@ function LoginScreen(): JSX.Element {
   );
 }
 
-export default LoginScreen;
+export default Login;
