+++ src/pages/main-empty/main-empty.tsx
@@ -1,11 +1,4 @@
-import PlaceCard from '../../components/place-card/place-card';
-
-type MainScreenProps = {
-  placesCount: number;
-}
-
-function MainScreen({placesCount}: MainScreenProps): JSX.Element {
-  const placesArray = [...Array(placesCount).keys()];
+function MainEmpty(): JSX.Element {
 
   return (
     <div className="page page--gray page--main">
@@ -38,7 +31,7 @@ function MainScreen({placesCount}: MainScreenProps): JSX.Element {
         </div>
       </header>
 
-      <main className="page__main page__main--index">
+      <main className="page__main page__main--index page__main--index-empty">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
@@ -59,7 +52,7 @@ function MainScreen({placesCount}: MainScreenProps): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item tabs__item--active">
+                <a className="locations__item-link tabs__item">
                   <span>Amsterdam</span>
                 </a>
               </li>
@@ -69,7 +62,7 @@ function MainScreen({placesCount}: MainScreenProps): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="#">
+                <a className="locations__item-link tabs__item tabs__item--active" href="#">
                   <span>Dusseldorf</span>
                 </a>
               </li>
@@ -77,36 +70,14 @@ function MainScreen({placesCount}: MainScreenProps): JSX.Element {
           </section>
         </div>
         <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">312 places to stay in Amsterdam</b>
-              <form className="places__sorting" action="#" method="get">
-                <span className="places__sorting-caption">Sort by</span>
-                <span className="places__sorting-type" tabIndex={0}>
-                  Popular
-                  <svg className="places__sorting-arrow" width="7" height="4">
-                    <use xlinkHref="#icon-arrow-select"></use>
-                  </svg>
-                </span>
-                <ul className="places__options places__options--custom places__options--opened">
-                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-                  <li className="places__option" tabIndex={0}>Price: low to high</li>
-                  <li className="places__option" tabIndex={0}>Price: high to low</li>
-                  <li className="places__option" tabIndex={0}>Top rated first</li>
-                </ul>
-              </form>
-              <div className="cities__places-list places__list tabs__content">
-                {
-                  placesArray.map((place) => (
-                    <PlaceCard key={`place-${place + 1}`} />
-                  ))
-                }
+          <div className="cities__places-container cities__places-container--empty container">
+            <section className="cities__no-places">
+              <div className="cities__status-wrapper tabs__content">
+                <b className="cities__status">No places to stay available</b>
+                <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
               </div>
             </section>
-            <div className="cities__right-section">
-              <section className="cities__map map"></section>
-            </div>
+            <div className="cities__right-section"></div>
           </div>
         </div>
       </main>
@@ -114,4 +85,4 @@ function MainScreen({placesCount}: MainScreenProps): JSX.Element {
   );
 }
 
-export default MainScreen;
+export default MainEmpty;
