+++ src/pages/main/main.tsx
@@ -1,4 +1,12 @@
-function MainEmptyScreen(): JSX.Element {
+import PlaceCard from '../../components/place-card/place-card';
+
+type MainProps = {
+  placesCount: number;
+  placesOnPage: number;
+}
+
+function Main({ placesCount, placesOnPage }: MainProps): JSX.Element {
+  const placesArray = [...Array(placesCount).keys()];
 
   return (
     <div className="page page--gray page--main">
@@ -31,7 +39,7 @@ function MainEmptyScreen(): JSX.Element {
         </div>
       </header>
 
-      <main className="page__main page__main--index page__main--index-empty">
+      <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
@@ -52,7 +60,7 @@ function MainEmptyScreen(): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item">
+                <a className="locations__item-link tabs__item tabs__item--active">
                   <span>Amsterdam</span>
                 </a>
               </li>
@@ -62,7 +70,7 @@ function MainEmptyScreen(): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item tabs__item--active" href="#">
+                <a className="locations__item-link tabs__item" href="#">
                   <span>Dusseldorf</span>
                 </a>
               </li>
@@ -70,14 +78,38 @@ function MainEmptyScreen(): JSX.Element {
           </section>
         </div>
         <div className="cities">
-          <div className="cities__places-container cities__places-container--empty container">
-            <section className="cities__no-places">
-              <div className="cities__status-wrapper tabs__content">
-                <b className="cities__status">No places to stay available</b>
-                <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
+          <div className="cities__places-container container">
+            <section className="cities__places places">
+              <h2 className="visually-hidden">Places</h2>
+              <b className="places__found">{ placesCount } places to stay in Amsterdam</b>
+              <form className="places__sorting" action="#" method="get">
+                <span className="places__sorting-caption">Sort by</span>
+                <span className="places__sorting-type" tabIndex={0}>
+                  Popular
+                  <svg className="places__sorting-arrow" width="7" height="4">
+                    <use xlinkHref="#icon-arrow-select"></use>
+                  </svg>
+                </span>
+                <ul className="places__options places__options--custom places__options--opened">
+                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
+                  <li className="places__option" tabIndex={0}>Price: low to high</li>
+                  <li className="places__option" tabIndex={0}>Price: high to low</li>
+                  <li className="places__option" tabIndex={0}>Top rated first</li>
+                </ul>
+              </form>
+              <div className="cities__places-list places__list tabs__content">
+                {
+                  placesArray.slice(0, placesOnPage).map((place) => (
+                    <PlaceCard
+                      key = {`place-${place + 1}`}
+                    />
+                  ))
+                }
               </div>
             </section>
-            <div className="cities__right-section"></div>
+            <div className="cities__right-section">
+              <section className="cities__map map"></section>
+            </div>
           </div>
         </div>
       </main>
@@ -85,4 +117,4 @@ function MainEmptyScreen(): JSX.Element {
   );
 }
 
-export default MainEmptyScreen;
+export default Main;
